requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
langdetect==1.0.9
psycopg2-binary==2.9.10
six==1.16.0
sqlparse==0.5.1
